server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-decouple